FlaskApp/requirements.txt
Flask==2.2.5
requests==2.31.0
requests-oauthlib==1.3.1
python-dotenv==1.0.1
mailchimp-transactional==1.0.50
